refactor 컨트롤러스캐너 역할 분리
